+++ R/Smisc-package.R
@@ -5,8 +5,8 @@
 ##' \tabular{ll}{
 ##' Package: \tab Smisc\cr
 ##' Type: \tab Package\cr
-##' Version: \tab 0.2.15\cr
-##' Date: \tab 2015-10-31\cr
+##' Version: \tab 0.2.17\cr
+##' Date: \tab 2016-02-01\cr
 ##' License: \tab LICENSE\cr
 ##' }
 ##'