+++ R/selectElements.R
@@ -3,16 +3,16 @@
 ##' Validate selected elements from a character vector using a variety of selection mechanisms:  logical, names, or
 ##' numerical indexes
 ##'
-##' This function is useful for selecting rows or columns from data frames, while providing informative error messages
-##' if the \code{elements} for selection are specified incorrectly.  Note that logical vectors (or binary vectors of 0's
-##' and 1's) are not recycled as usual: they must be the same length as \code{cVec}.
+##' This function is especially useful for selecting rows or columns from data frames, while providing informative error messages
+##' if the \code{elements} for selection are specified incorrectly.  
 ##'
 ##' @export
-##' @param elements elements to select from \code{cVec}.  Can be a logical vector, a vector of 1's and 0's (that is
-##' converted to Boolean), a vector of numeric indexes, or a character vector of column names.
-##' If \code{elements == NULL}, \code{NULL} is returned.
+##' 
+##' @param elements elements to select from \code{cVec}.  Can be a logical vector, a vector of numeric indexes,
+##' or a character vector of element names.  Note that logical vectors are not recycled as usual:
+##' they must be the same length as \code{cVec}. If \code{elements == NULL}, \code{NULL} is returned.
 ##'
-##' @param cVec A character vector from which to select elements
+##' @param cVec A character vector from which to select elements (such as row names or column names)
 ##'
 ##' @return A character vector with elements that were selected from \code{cVec} using \code{elements}
 ##'
@@ -25,8 +25,11 @@
 ##' selectElements(c("a", "c"), cnames)
 ##' selectElements(c(1, 3), cnames)
 ##' selectElements(c(TRUE, FALSE, TRUE, FALSE, FALSE), cnames)
-##' selectElements(c(1, 0, 1, 0, 0), cnames)
 ##'
+##' # Select the 1st, 3rd, and 1st columns
+##' selectElements(c("a", "c", "a"), cnames)
+##' selectElements(c(1, 3, 1), cnames)
+##' 
 ##' # If you don't want to select any of them
 ##' selectElements(NULL, cnames)
 
@@ -65,8 +68,6 @@ selectElements <- function(elements, cVec) {
 
   }
 
-  # For all other formats:
-
   # If it's Boolean
   if (is.logical(elements)) {
 
@@ -78,27 +79,10 @@ selectElements <- function(elements, cVec) {
   # If it's numeric
   else if (is.numeric(elements)) {
 
-    # If we have only 0's and 1's
-    if (all(sort(unique(elements))[1:2] == c(0, 1))) {
-
-      if (length(elements) != length(cVec)) {
-        stop("When 0's and 1's are provided for 'elements', the length of 'elements' must match the length of 'cVec'")
-      }
-
-      # Select the names
-      elements <- as.logical(elements)
-
-    }
-
-    # If we have non-binary numeric indexes
-    else {
-
-      # Otherwise, we should have indexes that are in the set 1:length(cVec)
-      if (!all(elements %in% 1:length(cVec))) {
-        stop("The following numeric indexes provided to 'elements' are outside the range of indexes for 'cVec': ",
-             paste(setdiff(elements, 1:length(cVec)), collapse = ", "))
-      }
-
+    # Otherwise, we should have indexes that are in the set 1:length(cVec)
+    if (!all(elements %in% 1:length(cVec))) {
+      stop("The following numeric indexes provided to 'elements' are outside the range of indexes for 'cVec': ",
+           paste(setdiff(elements, 1:length(cVec)), collapse = ", "))
     }
 
   } # If it's numeric
@@ -106,7 +90,7 @@ selectElements <- function(elements, cVec) {
   # If elements didn't conform
   else {
 
-    stop("'elements' must be a character, numeric, or boolean vector, or a vector of 0's and 1's to select column names")
+    stop("'elements' must be a character, numeric, or a logical (boolean) vector to select column names")
 
   }